requirements.txt
###############################################################################
# Minimal requirements for running the core MediQ-GAN training & preprocessing
# (based ONLY on imports in: mediq-gan.py, data_utils.py, data_processing.py)
# If you use evaluation scripts, extra tooling, or quantum backends beyond
# PennyLane Lightning, you may need to re-add packages that were removed.
###############################################################################

numpy==2.2.6                # arrays, math utilities
torch==2.9.0                # core deep learning
torchvision==0.24.0         # datasets, transforms, utils
matplotlib==3.10.7          # saving generated image grids
PennyLane==0.42.3           # quantum circuit interface
PennyLane_Lightning==0.42.0 # lightning.qubit device used
pandas==2.3.3               # CSV metadata processing
tqdm==4.67.1                # progress bars
medmnist==3.0.2             # RetinaMNIST dataset export
pillow==12.0.0              # image file I/O (PIL)
